% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_utilities.R
\name{abundance_df}
\alias{abundance_df}
\title{Return abundance long-formatted data.frame from phyloseq object.}
\usage{
abundance_df(phy, taxa = FALSE, sample = FALSE, id = character(0), vars = c())
}
\arguments{
\item{phy}{A phyloseq object}

\item{taxa}{FALSE/TRUE for taxonomic information. Default is FALSE.}

\item{sample}{FALSE/TRUE for sample metadata. Default is FALSE.}

\item{id}{sample or study identifier. Default is "abcno".}

\item{vars}{List of variables of interest from the metadata (sample data).}
}
\value{
Abundance data.frame in long format. Each row is a sample feature.
}
\description{
Return abundance long-formatted data.frame from phyloseq object.
}
\examples{
library(phyloseq)
data(GlobalPatterns)
ab <- abundance_df(GlobalPatterns, taxa = TRUE, sample = TRUE, vars = c("SampleType"))
head(ab)
}
