% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_utilities.R
\name{filter_phy}
\alias{filter_phy}
\title{Filter phyloseq object based on abundance and prevalence.}
\usage{
filter_phy(
  phy,
  abundance = NULL,
  prevalence = NULL,
  compositional = TRUE,
  tss = TRUE
)
}
\arguments{
\item{phy}{A phyloseq object}

\item{abundance}{Minimum abundance (in relative abundance) to retain. Default is NULL.}

\item{prevalence}{Minimum prevalence (in percent) to retain. Default is NULL.}

\item{compositional}{FALSE/TRUE if count table is relative (compositional). Default is TRUE. If FALSE, absolute counts are returned but filtering is done in relative abundances.}

\item{tss}{FALSE/TRUE if relative abundance is re-calculated after filtering taxa. Default is TRUE.}
}
\value{
A filtered phyloseq object.
}
\description{
Filter phyloseq object based on abundance and prevalence.
}
\examples{
library(phyloseq)
data(GlobalPatterns)
GlobalPatterns
filter_phy(GlobalPatterns, abundance = 0.01, prevalence = 5)
}
