% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_utilities.R
\name{get_otu}
\alias{get_otu}
\title{Return OTU table from phyloseq object as data.frame}
\usage{
get_otu(phy)
}
\arguments{
\item{phy}{A phyloseq object.}
}
\value{
a data.frame containing abundances with samples (rows) and taxa(columns).
}
\description{
Return OTU table from phyloseq object as data.frame
}
\examples{
library(phyloseq)
data(GlobalPatterns)
otumat <- get_otu(GlobalPatterns)
head(otumat)[,1:5]
}
