% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_utilities.R
\name{get_virome_host_phy}
\alias{get_virome_host_phy}
\title{Return virome phyloseq object formatted ready for agglomerating bacteria host taxa.}
\usage{
get_virome_host_phy(phy)
}
\arguments{
\item{phy}{A virome phyloseq object with host information.}
}
\value{
A virome phyloseq object with taxonomic information related to host, i.e. easy to agglomerate.
}
\description{
Return virome phyloseq object formatted ready for agglomerating bacteria host taxa.
}
