% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweetmap.R
\name{lo}
\alias{lo}
\title{Pretty correlation plots}
\usage{
lo(
  rown,
  coln,
  nrow,
  ncol,
  cellheight = NA,
  cellwidth = NA,
  treeheight_col,
  treeheight_row,
  legend,
  annotation_row,
  annotation_col,
  annotation_colors,
  annotation_legend,
  annotation_names_row,
  annotation_names_col,
  main,
  fontsize,
  fontsize_row,
  fontsize_col,
  angle_col,
  gaps_row,
  gaps_col,
  ...
)
}
\description{
Implementation of pheatmap for correlation plots
}
