% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_utilities.R
\name{make_mradat}
\alias{make_mradat}
\title{Make a data.frame with handy taxon-wise summaries}
\usage{
make_mradat(phy)
}
\arguments{
\item{phy}{A phyloseq object}
}
\value{
a data.frame which contains the folowing variables: tax (taxa names), mra (mean relative abundance), prevalence, and a copy of the tax table appended.
}
\description{
Make a data.frame with handy taxon-wise summaries
}
\examples{
library(phyloseq)
data(GlobalPatterns)
mradat <- make_mradat(GlobalPatterns)
head(mradat)
}
