% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_utilities.R
\name{na_phy}
\alias{na_phy}
\title{Return number of NAs in sample data}
\usage{
na_phy(phy)
}
\arguments{
\item{phy}{A phyloseq object}
}
\value{
returns a named vector of the number of NAs in sample variables
}
\description{
Return number of NAs in sample data
}
\examples{
library(phyloseq)
data(GlobalPatterns)
na_phy(GlobalPatterns)
}
