% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_utilities.R
\name{otu_df}
\alias{otu_df}
\title{Return OTU table from phyloseq object as data.frame}
\usage{
otu_df(phy)
}
\arguments{
\item{phy}{A phyloseq object.}
}
\value{
a data.frame containing abundances with samples (rows) and taxa(columns).
}
\description{
Return OTU table from phyloseq object as data.frame
}
\examples{
library(phyloseq)
data(GlobalPatterns)
otumat <- otu_df(GlobalPatterns)
head(otumat)[,1:5]
}
