% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{roundex}
\alias{roundex}
\title{Round to exact number of digits
Useful when plotting, since round() cuts away zeroes at the end}
\usage{
roundex(vec, n)
}
\arguments{
\item{vec}{Numeric vector that you want to round}

\item{n}{number of decimal places}
}
\value{
Character vector with exact rounding.
}
\description{
Round to exact number of digits
Useful when plotting, since round() cuts away zeroes at the end
}
\examples{
round(3.9999, 3)
roundex(3.9999, 3)
}
