% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_utilities.R
\name{sample_df}
\alias{sample_df}
\title{Return sample data from phyloseq object as data.frame.}
\usage{
sample_df(phy)
}
\arguments{
\item{phy}{A phyloseq object.}
}
\value{
a data.frame containing sample data with samples (rows) and metadata (columns).
}
\description{
Return sample data from phyloseq object as data.frame.
}
\examples{
library(phyloseq)
data(GlobalPatterns)
samdat <- sample_df(GlobalPatterns)
head(samdat)[,1:5]
}
