% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{subset_dist}
\alias{subset_dist}
\title{Remove samples from distance object based on the samples of a phyloseq object}
\usage{
subset_dist(d, phy)
}
\arguments{
\item{d}{A numeric matrix, data frame or "dist" object.}

\item{phy}{A phyloseq object.}
}
\value{
A subsetted "dist" object.
}
\description{
Remove samples from distance object based on the samples of a phyloseq object
}
\examples{
library(phyloseq)
data(GlobalPatterns)
dist.bray <- distance(GlobalPatterns, "bray")
subset_dist(dist.bray, subset_samples(GlobalPatterns, SampleType == "Feces"))
}
