% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_utilities.R
\name{tax_df}
\alias{tax_df}
\title{Return tax data from phyloseq object as data.frame}
\usage{
tax_df(phy)
}
\arguments{
\item{phy}{A phyloseq object.}
}
\value{
a data.frame containing tax data with taxa (rows) and ranks (columns). If the phyloseq object is
agglomerated to lower taxonomic resolution, the lower ranks are not shown.
}
\description{
Return tax data from phyloseq object as data.frame
}
\examples{
library(phyloseq)
data(GlobalPatterns)
taxmat <- tax_df(GlobalPatterns)
head(taxmat)
}
