% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{tidylog}
\alias{tidylog}
\title{Tidy function for logistic regression
Adds n and n_missing}
\usage{
tidylog(model, ...)
}
\arguments{
\item{model}{The glm model object}

\item{...}{Other arguments passed on to tidy()}
}
\value{
A data.frame with model parameters, similar to tidy()
}
\description{
Tidy function for logistic regression
Adds n and n_missing
}
\examples{
library(tidyverse)
glm(factor(am) ~ disp, data = mtcars, family = binomial) \%>\% tidylog(exp = TRUE, conf.int = TRUE)

}
