% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_utilities.R
\name{transform_phy}
\alias{transform_phy}
\title{Common transformations for phyloseq count data.}
\usage{
transform_phy(phy, transform = "compositional", pseudocount = NULL)
}
\arguments{
\item{phy}{A phyloseq object.}

\item{transform}{Transformation to be applied: "compositional" (i.e. relative abundance), "clr" (centered log ratio), "log" (log10) and "pa" (presence/absence).
By default "clr" applies a pseudocount of min(relative abundance)/2 and "log" applies a pseudocount of min/2 before taking logs. The user can also define a certain value
for the pseudocount.}

\item{pseudocount}{Pseudocount to be used in "clr" or "log" transformations. Default is NULL}
}
\value{
Transformed phyloseq object
}
\description{
Common transformations for phyloseq count data.
}
\examples{
library(phyloseq)
data(esophagus)

# No transformation
head(otu_df(esophagus))

# Relative abundance
head(otu_df(transform_phy(esophagus, transform = "compositional")))

# Centered log-ratio (clr) transformation
head(otu_df(transform_phy(esophagus, transform = "clr")))
head(otu_df(transform_phy(esophagus, transform = "clr", pseudocount = 0.1)))

# Log10 transformation
head(otu_df(transform_phy(esophagus, transform = "log")))
head(otu_df(transform_phy(esophagus, transform = "log", pseudocount = 1)))

# Presence/absence transformation
head(otu_df(transform_phy(esophagus, transform = "log")))
}
