% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utilities.R
\name{relabu}
\alias{relabu}
\title{Compute relative abundances}
\usage{
relabu(x)
}
\arguments{
\item{x}{A numeric vector of counts}
}
\value{
A numeric vector of relative abundances
}
\description{
Compute relative abundances
}
\examples{
transform_sample_counts(phy, relabu)
}
